+++ src/configUI.unit.test.ts
@@ -63,6 +63,7 @@ describe('check package ruler', () => {
                 ['AhkNekoHelp.CodeLens.showClassReference', 'boolean'],
                 ['AhkNekoHelp.CodeLens.showFuncReference', 'boolean'],
                 ['AhkNekoHelp.CodeLens.showLabelReference', 'boolean'],
+                ['AhkNekoHelp.CodeLens.showComObjConnectRegisterStrReference', 'boolean'],
                 ['AhkNekoHelp.CodeLens.showDevTool', 'boolean'],
                 ['AhkNekoHelp.CodeLens.showFileReport', 'boolean'],
                 // [1]